1.15-pre6

+++ net/minecraft/nbt/ByteArrayTag.java
@@ -26,7 +26,7 @@ extends CollectionTag<ByteTag> {
         public ByteArrayTag load(DataInput dataInput, int i, NbtAccounter nbtAccounter) throws IOException {
             nbtAccounter.accountBits(192L);
             int j = dataInput.readInt();
-            nbtAccounter.accountBits(8 * j);
+            nbtAccounter.accountBits(8L * (long)j);
             byte[] bs = new byte[j];
             dataInput.readFully(bs);
             return new ByteArrayTag(bs);
