+++ net/minecraft/nbt/IntArrayTag.java
@@ -26,7 +26,7 @@ extends CollectionTag<IntTag> {
         public IntArrayTag load(DataInput dataInput, int i, NbtAccounter nbtAccounter) throws IOException {
             nbtAccounter.accountBits(192L);
             int j = dataInput.readInt();
-            nbtAccounter.accountBits(32 * j);
+            nbtAccounter.accountBits(32L * (long)j);
             int[] is = new int[j];
             for (int k = 0; k < j; ++k) {
                 is[k] = dataInput.readInt();
