+++ net/minecraft/nbt/LongArrayTag.java
@@ -27,7 +27,7 @@ extends CollectionTag<LongTag> {
         public LongArrayTag load(DataInput dataInput, int i, NbtAccounter nbtAccounter) throws IOException {
             nbtAccounter.accountBits(192L);
             int j = dataInput.readInt();
-            nbtAccounter.accountBits(64 * j);
+            nbtAccounter.accountBits(64L * (long)j);
             long[] ls = new long[j];
             for (int k = 0; k < j; ++k) {
                 ls[k] = dataInput.readLong();
